requirements.txt
attrs==21.4.0
discord.py==2.0.1
lark==1.1.2
aiohttp==3.8.3